% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBICglasso_wi.R
\name{EBICglasso_wi}
\alias{EBICglasso_wi}
\title{Estimate the precision matrix}
\usage{
EBICglasso_wi(S, n, gamma = 0.5, penalize.diagonal = FALSE,
  nlambda = 100, lambda.min.ratio = 0.01, returnAllResults = FALSE,
  checkPD = TRUE, penalizeMatrix, countDiagonal = FALSE,
  refit = FALSE, ...)
}
\arguments{
\item{S}{A covariance or correlation matrix}

\item{n}{Sample size used in computing \code{S}}

\item{gamma}{EBIC tuning parameter. 0.5 is generally a good choice.
Setting to zero will cause regular BIC to be used.}

\item{penalize.diagonal}{Should the diagonal be penalized?}

\item{nlambda}{Number of lambda values to test.}

\item{lambda.min.ratio}{Ratio of lowest lambda value compared to maximal lambda}

\item{returnAllResults}{If \code{TRUE} this function does not return a network
but the results of the entire glasso path.}

\item{checkPD}{If \code{TRUE}, the function will check if \code{S} is
positive definite and return an error if not. It is not advised to use
 a non-positive definite matrix as input as (a) that can not be a covariance
 matrix and (b) glasso can hang if the input is not positive definite.}

\item{penalizeMatrix}{Optional logical matrix to indicate which elements are penalized}

\item{countDiagonal}{Should diagonal be counted in EBIC computation? Defaults to \code{FALSE}.
Set to \code{TRUE} to mimic qgraph < 1.3 behavior (not recommended!).}

\item{refit}{Logical, should the optimal graph be refitted without LASSO regularization?
Defaults to \code{FALSE}.}

\item{...}{functions to be passed to \link[glasso]{glasso}}
}
\value{
a \eqn{p \times p} estimated precision matrix
}
\description{
A function to compute a sparse gaussian graphical model. This function is
    taken directly from the \link[qgraph]{qgraph} package function \link[qgraph]{EBICglasso}
    version 1.4.4 (Epskamp et al. 2012).
    The only alteration is to remove the wi2net function call so as to output
    the precision matrix directly, rather than the standardised partial correlations
}
\examples{
data(riskcor)
omega <- EBICglasso_wi(riskcor,n=69)

# Results can be plotted as a GGM
parcor <- qgraph::wi2net(omega)
pos <- matrix(c(2,0,-2,-1,-2,1,0,2,0.5,0,0,-2),6,2,byrow=TRUE)
qgraph::qgraph(parcor, labels=rownames(riskcor), layout=pos,
repulsion=.8, vsize=c(10,15), theme="colorblind")
}
\references{
\insertRef{qgraph}{SEset}

\insertRef{glasso}{SEset}

\insertRef{friedman2008sparse}{SEset}

\insertRef{Matrix}{SEset}

\insertRef{ryan2019}{SEset}

\insertRef{hoorelbeke2016interplay}{SEset}
}
