% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parents.R
\name{find_parents}
\alias{find_parents}
\title{Return parent indices from a (weighted) DAG for a given child}
\usage{
find_parents(mat, child)
}
\arguments{
\item{mat}{An \eqn{p \times p} weights or adjacency matrix}

\item{child}{Index giving the position of the child node}
}
\value{
a vector containing index numbers defining the parent nodes
}
\description{
Return parent indices from a (weighted) DAG for a given child
}
\references{
\insertRef{ryan2019}{SEset}
}
\seealso{
\code{\link{r2_distribution}}
}
