% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psi.R
\name{get_psi}
\alias{get_psi}
\title{Calculate residual-covariance matrix based on a path model and covariance matrix}
\usage{
get_psi(B, sigma, digits = 3)
}
\arguments{
\item{B}{input \eqn{p \times p} linear SEM weights matrix}

\item{sigma}{variance-covariance matrix of the variables}

\item{digits}{how many digits to round the result to}
}
\value{
a \eqn{p \times p} residual variance-covariance matrix
}
\description{
Takes an ordered path model and corresponding variance-covariance matrix and computes the appropriate
residual covariance matrix (psi)
}
