% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder2.R
\name{reorder2}
\alias{reorder2}
\title{Re-order rows and columns}
\usage{
reorder2(matrix, names)
}
\arguments{
\item{matrix}{input matrix to be re-arranged. Must have rows and columns named}

\item{names}{character vector containing the dimension names of the input matrix
in the desired ordering}
}
\value{
input matrix with rows and columns sorted according to names
}
\description{
Takes a matrix and re-orders the rows and columns to some target ordering
}
\examples{
data(riskcor)

# first define an ordered vector of names
names <- rownames(riskcor)
names <- c(names[1],names[2],names[3],names[4],names[6],names[5])

reorder2(riskcor,names)

# The fifth and sixth row and column have been switched
print(riskcor)
}
\references{
\insertRef{memisc}{SEset}
}
\seealso{
\code{\link{order_gen}}, \code{\link{precision_to_SEset}}
}
